Move find_integer into junk drawer
